New tab for metriq-gym results
